Refactor sweepers namespace for clarity and maintainability
This commit introduces significant refactoring in the `BowlingMegabucks.TournamentManager.Sweepers` namespace, focusing on the addition, retrieval, and presentation of sweepers. Key changes include:

- Added detailed XML documentation for improved readability.
- Restructured interfaces and classes for better separation of concerns.
- Implemented error handling in presenter classes for user feedback.
- Enhanced view models with new properties and methods for data encapsulation.
- Integrated dependency injection for improved service management.
- Removed unused code to streamline the codebase.
- Added extension methods for converting between view models and domain models.
