--- source/main/Sweepers/Add/IAddSweeperView.cs
@@ -1,22 +1,0 @@
-﻿
-namespace BowlingMegabucks.TournamentManager.Sweepers.Add;
-internal interface IView
-{
-    IViewModel Sweeper { get; }
-
-    void DisplayError(string message);
-
-    void DisplayMessage(string message);
-
-    void Disable();
-
-    void KeepOpen();
-
-    void BindDivisions(IEnumerable<Divisions.IViewModel> divisions);
-
-    TournamentId TournamentId { get; }
-
-    bool IsValid();
-
-    void Close();
-}
