--- source/main/Sweepers/Results/ISweeperResultsView.cs
@@ -1,8 +1,0 @@
-﻿
-namespace BowlingMegabucks.TournamentManager.Sweepers.Results;
-internal interface IView
-{
-    void DisplayError(string message);
-
-    void BindResults(ICollection<IViewModel> results);
-}
