--- source/main/Sweepers/Retrieve/IRetrieveSweepersView.cs
@@ -1,16 +1,0 @@
-﻿
-namespace BowlingMegabucks.TournamentManager.Sweepers.Retrieve;
-internal interface IView
-{
-    TournamentId TournamentId { get; }
-
-    void BindSweepers(IEnumerable<IViewModel> squads);
-
-    void Disable();
-
-    void DisplayError(string message);
-
-    SquadId? AddSweeper(TournamentId tournamentId);
-
-    Task RefreshSweepersAsync(CancellationToken cancellationToken);
-}
