+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Sweepers/Complete/CompleteSweeperAdapter.cs
