+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Sweepers/Portal/SweeperPortalPresenter.cs
@@ -1,7 +1,11 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
 namespace BowlingMegabucks.TournamentManager.Sweepers.Portal;
-internal class Presenter
+
+/// <summary>
+/// Handles the presentation logic for the sweeper portal, including loading and completing sweepers.
+/// </summary>
+public class Presenter
 {
     private readonly IView _view;
     private readonly Retrieve.IAdapter _retrieveSquadAdapter;
@@ -9,10 +13,14 @@ internal class Presenter
     private readonly Lazy<Complete.IAdapter> _completeSweeperAdapter;
     private Complete.IAdapter CompleteSweeperAdapter => _completeSweeperAdapter.Value;
 
+    /// <summary>
+    /// Initializes a new instance of the <see cref="Presenter"/> class.
+    /// </summary>
+    /// <param name="view">The view interface for the sweeper portal.</param>
+    /// <param name="services">The service provider for dependency injection.</param>
     public Presenter(IView view, IServiceProvider services)
     {
         _view = view;
-
         _retrieveSquadAdapter = services.GetRequiredService<Retrieve.IAdapter>();
         _completeSweeperAdapter = new Lazy<Complete.IAdapter>(services.GetRequiredService<Complete.IAdapter>);
     }
@@ -20,9 +28,9 @@ internal class Presenter
     /// <summary>
     /// Unit Test Constructor
     /// </summary>
-    /// <param name="mockView"></param>
-    /// <param name="mockRetrieveSquadAdapter"></param>
-    /// <param name="mockCompleteSweeperAdapter"></param>
+    /// <param name="mockView">The mock view for testing.</param>
+    /// <param name="mockRetrieveSquadAdapter">The mock retrieve squad adapter for testing.</param>
+    /// <param name="mockCompleteSweeperAdapter">The mock complete sweeper adapter for testing.</param>
     internal Presenter(IView mockView, Retrieve.IAdapter mockRetrieveSquadAdapter, Complete.IAdapter mockCompleteSweeperAdapter)
     {
         _view = mockView;
@@ -30,6 +38,14 @@ internal class Presenter
         _completeSweeperAdapter = new Lazy<Complete.IAdapter>(() => mockCompleteSweeperAdapter);
     }
 
+    /// <summary>
+    /// Loads sweeper details asynchronously and updates the view.
+    /// </summary>
+    /// <param name="cancellationToken">A cancellation token to observe while waiting for the task to complete.</param>
+    /// <returns>A task representing the asynchronous operation.</returns>
+    /// <remarks>
+    /// This method retrieves sweeper details, handles errors, and binds the results to the view.
+    /// </remarks>
     public async Task LoadAsync(CancellationToken cancellationToken)
     {
         var squad = await _retrieveSquadAdapter.ExecuteAsync(_view.Id, cancellationToken).ConfigureAwait(true);
@@ -37,21 +53,26 @@ internal class Presenter
         if (_retrieveSquadAdapter.Error != null)
         {
             _view.DisplayError(_retrieveSquadAdapter.Error.Message);
-
             _view.Close();
-
             return;
         }
 
         _view.SetPortalTitle($"{squad!.Date:MM/dd/yyyy hh:mmtt}");
-
         _view.SetStartingLane(squad.StartingLane);
         _view.SetNumberOfLanes(squad.NumberOfLanes);
         _view.SetMaxPerPair(squad.MaxPerPair);
         _view.SetComplete(squad.Complete);
     }
 
-    internal async Task CompleteAsync(CancellationToken cancellationToken)
+    /// <summary>
+    /// Completes the sweeper asynchronously and updates the view.
+    /// </summary>
+    /// <param name="cancellationToken">A cancellation token to observe while waiting for the task to complete.</param>
+    /// <returns>A task representing the asynchronous operation.</returns>
+    /// <remarks>
+    /// This method prompts for confirmation, completes the sweeper, handles errors, and updates the view accordingly.
+    /// </remarks>
+    public async Task CompleteAsync(CancellationToken cancellationToken)
     {
         if (!_view.Confirm("Are you sure you want to complete this sweeper?"))
         {
