+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Sweepers/Results/SweeperResultsAdapter.cs
